requirements.txt
altgraph==0.17.4
asttokens==2.4.1
auto-py-to-exe==2.44.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.50.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.0
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pybliometrics==3.6
pycparser==2.22
Pygments==2.17.2
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.4.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
tzdata==2023.4
urllib3==2.2.1
wcwidth==0.2.13
whichcraft==0.6.1
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.4.post2
